+++ dutch/app/lib/game-actions.js
@@ -131,13 +131,15 @@ function createGameActions(deps) {
     }
     round.throwIn.open = false;
     player.cards.splice(index, 1);
+    const article = ['8', 'A'].includes(card.rank) ? 'an' : 'a';
     deps.pushDiscard(card, player.id, 'threw in', {
       allowThrowIn: false,
       observationSource: 'throw-in',
       observationActorId: player.id,
       removedSlotOwnerId: player.id,
       removedSlotIndex: index,
-      removedSlotSource: 'throw-in'
+      removedSlotSource: 'throw-in',
+      infoEventText: player.name + ' threw in ' + article + ' ' + deps.label(card)
     });
     deps.highlightPileForAll('event', 3000);
     return { valid: true, card, index };
@@ -157,6 +159,7 @@ function createGameActions(deps) {
       deps.markHandCardChanged(target.id, card.id);
       deps.observeAceForAllBots(player.id, target.id);
       deps.addLog(player.name + ' gave a card to ' + target.name);
+      deps.showInfoEvent(player.name + ' used Ace add');
     }
     deps.finishSpecial();
     return true;
@@ -176,6 +179,7 @@ function createGameActions(deps) {
     }
     if (deps.observeDecisionForAllBots) deps.observeDecisionForAllBots(player.id, 'queen-target', { targetId: target.player.id });
     deps.addLog(player.name + ' used Queen peek');
+    deps.showInfoEvent(player.name + ' used Queen peek');
     deps.finishSpecial();
     return true;
   }