+++ dutch/app/lib/game-services.js
@@ -225,6 +225,7 @@ function createGameServices(options) {
     revealCardTo,
     highlightCardForAll,
     highlightPileForAll,
+    showInfoEvent,
     markHandCardChanged,
     clearHandHighlightsForPlayer
   } = cardFlow;
@@ -244,7 +245,8 @@ function createGameServices(options) {
     findPlayer,
     currentPlayer,
     specialName,
-    advanceTurn
+    advanceTurn,
+    showInfoEvent
   });
 
   function updateStageAfterQueue() {
@@ -288,6 +290,7 @@ function createGameServices(options) {
     addLog,
     removeSlotForAllBots,
     highlightPileForAll,
+    showInfoEvent,
     findPlayer,
     isProtectedSpecialTarget,
     observeAceForAllBots,